+++ help50/internal.py
@@ -1,0 +1,81 @@
+import contextlib
+import re
+import sys
+
+import lib50
+
+from . import HELPERS, PREPROCESSORS
+
+lib50.LOCAL_PATH = "~/.local/share/help50"
+
+
+@contextlib.contextmanager
+def _syspath(newpath):
+    """ Useful contextmanager that temporarily replaces sys.path. """
+    oldpath = sys.path
+    sys.path = newpath
+    try:
+        yield
+    finally:
+        sys.path = oldpath
+
+
+def load_config(dir):
+    """ Read cs50 YAML file and apply default configuration to unspecified fields.  """
+    options = { "helpers": ["helpers"] }
+    try:
+        config_file = lib50.config.get_config_filepath(dir)
+
+        with open(config_file) as f:
+            config = lib50.config.Loader("help50").load(f.read())
+    except lib50.InvalidConfigError:
+        raise Error("Failed to parse help50 config, please let sysadmins@cs50.harvard.edu know!")
+
+
+    if isinstance(config, dict):
+        options.update(config)
+
+    if isinstance(options["helpers"], str):
+        options["helpers"] = [options["helpers"]]
+
+    return options
+
+
+def load_helpers(slug):
+    """ Download helpers to a local directory via lib50. """
+    try:
+        helpers_dir = lib50.local(slug)
+    except lib50.Error:
+        raise Error("Failed to fetch helpers, please ensure that you are connected to the internet!")
+
+    config = load_config(helpers_dir)
+    for helper in config["helpers"]:
+        with _syspath([str(helpers_dir)]):
+            try:
+                __import__(helper)
+            except ImportError:
+                raise Error("Failed to load helpers, please let sysadmins@cs50.harvard.edu know!")
+
+
+def get_help(output):
+    """
+    Given an error message, try every helper registered with help50 and return the output of the
+    first one that matches.
+    """
+    # Strip ANSI codes
+    output = re.sub(r"\x1B\[[0-?]*[ -/]*[@-~]", "", output)
+
+    for domain in HELPERS.keys():
+        processed = output
+        for pre in PREPROCESSORS.get(domain, []):
+            processed = pre(processed)
+        lines = processed.splitlines()
+        for i in range(len(lines)):
+            slice = lines[i:]
+            for helper in HELPERS.get(domain, []):
+                try:
+                    before, after = helper(slice)
+                except TypeError:
+                    pass
+                else:
+                    return before, " ".join(after)
